backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gTTS==2.3.2
requests==2.31.0
google-generativeai
python-dotenv
